js/src/xpconnect/tests/mochitest/test_bug517163.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=517163
-->
<head>
  <title>Test for Bug 517163</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=517163">Mozilla Bug 517163</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 517163 **/
netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var rval = Components.utils.lookupMethod(document, "getElementById")("content");
ok(/XPCNativeWrapper/(rval.toString()), "got an XPCNativeWrapper back");

</script>
</pre>
</body>
</html>
